restructuring db package
