--- internal/database/seeds/users.json
@@ -1,9 +1,0 @@
-{
-  "users": [
-    {
-      "username": "jrico",
-      "password": "localdev",
-      "passwordConfirmation": "localdev"
-    }
-  ]
-}
