+++ internal/db/postgresql_db/database.go
@@ -1,4 +1,4 @@
-package postgresql
+package postgresql_db
 
 import (
 	"context"
