+++ internal/db/redis_db/database.go
@@ -1,4 +1,4 @@
-package redis
+package redis_db
 
 import (
 	"context"
